First attempt at reuseable txt data processing tools
calc_and_save_bodes.py is a first attempt at a module for converting
time domain swept sine data sets to Bodes.  bode_opts.txt is a file
that is parsed by calc_and_save_bodes to determine how to create the
bodes and handle the data.

Similarly, log_downsample_bodes.py is a module (mainly a script) for
log downsampling of Bode data and log_opts.txt is a file containing
options to be parsed.

The ideas is that bode_opts.txt and log_opts.txt files would be placed
in each directory containing Bode data and then calc_and_save_bodes
and log_downsample_bodes would be called passing the fullpaths to the
txt options files.