Periodic check-in for SFLR automated ROM.
Working toward OL TMM contour plotting to find all poles and zeros in
a region of the s-plane.  The added or modified files help me find and
verify a Maxima-based Python module for finding the OL Bodes of the
SFLR.  This module will be used to calculate the contour plots.